requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
discord-py-slash-command==4.2.1
discord.py @ git+https://github.com/Rapptz/discord.py@59f877fcf013c4ddeeb2b39fc21f03e76f995461
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
jiter==0.6.0
multidict==6.0.5
mutagen==1.47.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
yt-dlp==2024.4.9
